core: fix new signature for generators

+++ inc/kos_utils.h
@@ -190,7 +190,7 @@ int KOS_print_to_cstr_vec(KOS_CONTEXT             ctx,
                           KOS_QUOTE_STR           quote_str,
                           struct KOS_VECTOR_S    *cstr_vec,
                           const char             *sep,
-                          unsigned                sep_len);
+                          size_t                  sep_len);
 
 KOS_API
 int KOS_append_cstr(KOS_CONTEXT          ctx,
